+++ bindings/nodejs/CHANGELOG.md
@@ -21,6 +21,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 2.0.0-alpha.2 - 2024-MM-DD
 
+### Added
+
+- `Utils::blockWorkScore()`;
+
 ### Fixed
 
 - `RestrictedAddress` discriminator usage;